Extract and test manifest customization and ENV parsing
