+++ .eslintrc.js
@@ -101,7 +101,7 @@ module.exports = {
   },
   overrides: [
     {
-      files: ["webpack.*.js", "*.config.js"],
+      files: ["webpack.*.js", "*.config.js", "scripts/*"],
       // Full config: https://github.com/pixiebrix/eslint-config-pixiebrix/blob/main/development.js
       extends: ["pixiebrix/development"],
       rules: {
@@ -110,6 +110,7 @@ module.exports = {
     },
     {
       files: [
+        "*/scripts/*",
         "**/__mocks__/**",
         "**/testUtils/**",
         "**/*.test.ts",
@@ -129,6 +130,16 @@ module.exports = {
       rules: {
         "testing-library/render-result-naming-convention": "off",
         "testing-library/no-await-sync-queries": "off",
+      },
+    },
+    {
+      // These rules should not be enabled for JS files
+      files: ["**/*.js"],
+      rules: {
+        "@typescript-eslint/no-unsafe-member-access": "off",
+        "@typescript-eslint/no-unsafe-call": "off",
+        "@typescript-eslint/no-unsafe-assignment": "off",
+        "@typescript-eslint/no-unsafe-return": "off",
       },
     },
   ],
