+++ scripts/manifest.test.js
@@ -1,0 +1,55 @@
+/*
+ * Copyright (C) 2023 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @shopify/jest/no-snapshots -- We want to specifically commit the entire customized manifest as a snapshot */
+/* eslint-disable no-restricted-imports -- Aliases don't work outside built files */
+
+import manifest from "../src/manifest.json";
+import { loadEnv } from "./env";
+import customizeManifest from "./manifest";
+
+loadEnv();
+
+function dropVariableProperties(manifest) {
+  delete manifest.version;
+  delete manifest.version_name;
+  delete manifest.key;
+  return manifest;
+}
+
+describe("customizeManifest", () => {
+  test("mv2", () => {
+    expect(
+      dropVariableProperties(
+        customizeManifest(manifest, {
+          env: process.env,
+        }),
+      ),
+    ).toMatchSnapshot();
+  });
+  test("mv3", () => {
+    expect(
+      dropVariableProperties(
+        customizeManifest(manifest, {
+          env: process.env,
+          manifestVersion: 3,
+        }),
+      ),
+    ).toMatchSnapshot();
+  });
+});
